split gn_commons models in two files
