+++ backend/geonature/core/gn_commons/models/__init__.py
@@ -1,0 +1,2 @@
+from .base import *
+from .additional_fields import *
